// src/app/Models/ServiceSpeed.ts
export class ServiceSpeed {
    id: number=0;
    creationDate?: Date;
    modificationDate?: Date;
    createdBy?: any;
    userGroup?: any;
    modifyiedBy?: any;   
    createdByTeam?: any;
    modifyiedByTeam?: any;
    value?: any;
    orderInList?: any;
    serviceTypeID?: any;
    serviceType?: any;
}